trying to fix timeout of mac builds with caching

+++ getosxdeps.sh
@@ -1,18 +1,26 @@
-#/bin/sh
+#!/bin/sh
 #load Vespucci dependencies for osx using homebrew
 brew tap homebrew/science
-brew install qt5
-brew install hdf5 --with-unsupported
-brew install libxml2
-brew install arpack
-brew install superlu43
-brew install wget
-brew link qt5
-brew link hdf5
-brew link libxml2
-brew link arpack
-brew link --force superlu43
-brew link wget
+if [ ! -f /usr/local/lib/libhdf5.a]; then
+    brew install hdf5 --with-unsupported
+    brew link hdf5
+fi
+if [ ! -f /usr/local/lib/QtCore.framework ]; then
+    brew install qt5
+    brew link qt5
+fi
+if [ ! -f /usr/local/lib/libarpack.dylib ]; then
+    brew install arpack
+    brew link arpack
+fi
+if [! -f /usr/local/lib/libxml2.dylib ]; then
+    brew install libxml2
+    brew link libxml2
+fi
+if [! -f /usr/local/bin/wget ]; then
+    brew install wget
+    brew link wget
+fi
 wget https://github.com/VespucciProject/Vespucci_dependencies/releases/download/1/Vespucci_dependencies_macOS.zip
 unzip Vespucci_dependencies_macOS.zip
